+++ src/clientMain.ts
@@ -2,7 +2,7 @@
 
 import * as fs from "fs";
 import * as path from "path";
-import { commands, DiagnosticCollection, ExtensionContext, window, workspace } from "vscode";
+import { commands, DiagnosticCollection, ExtensionContext, Uri, window, workspace } from "vscode";
 import { absolutePath } from "./helpers/AbsolutePath";
 import * as tools from "./toolchain/SwiftTools";
 import * as config from "./vscode/config-helpers";
@@ -18,7 +18,7 @@ export let diagnosticCollection: DiagnosticCollection;
 
 let mostRecentRunTarget = "";
 
-export function activate(context: ExtensionContext) {
+export async function activate(context: ExtensionContext) {
   output.init(context);
 
   if (workspace.getConfiguration().get<boolean>("sde.enable") === false) {
@@ -28,8 +28,8 @@ export function activate(context: ExtensionContext) {
   tools.setRunning(false);
   output.build.log("Activating SDE");
 
-  initConfig();
-  lsp.startLSPClient(context);
+  await initConfig();
+  await lsp.startLSPClient(context);
 
   //commands
   let toolchain = new tools.Toolchain(
@@ -41,8 +41,9 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(toolchain.start());
   context.subscriptions.push(
     commands.registerCommand("sde.commands.build", () => toolchain.build()),
-    commands.registerCommand("sde.commands.restartLanguageServer", () =>
-      lsp.restartLSPClient(context)
+    commands.registerCommand(
+      "sde.commands.restartLanguageServer",
+      async () => await lsp.restartLSPClient(context)
     ),
     commands.registerCommand("sde.commands.run", () => toolchain.runStart()),
     commands.registerCommand("sde.commands.selectRun", () => {
@@ -75,9 +76,13 @@ export function activate(context: ExtensionContext) {
   );
 
   // respond to settings changes
-  workspace.onDidChangeConfiguration(event => {
-    if (event.affectsConfiguration("sde")) {
-      // reload things as necessary
+  workspace.onDidChangeConfiguration(async event => {
+    if (
+      event.affectsConfiguration("sde") ||
+      event.affectsConfiguration("swift") ||
+      event.affectsConfiguration("sourcekit-lsp")
+    ) {
+      await lsp.restartLSPClient(context);
     }
   });
 
@@ -85,11 +90,11 @@ export function activate(context: ExtensionContext) {
   toolchain.build();
 }
 
-function initConfig() {
+async function initConfig() {
   checkToolsAvailability();
 }
 
-function checkToolsAvailability() {
+async function checkToolsAvailability() {
   swiftBinPath = absolutePath(workspace.getConfiguration().get("swift.path.swift_driver_bin"));
   swiftBuildParams = <string[]>workspace.getConfiguration().get("sde.swiftBuildingParams") || [
     "build",
@@ -102,21 +107,29 @@ function checkToolsAvailability() {
   skProtocolProcess = sourcekitePath;
   skProtocolProcessAsShellCmd = sourcekitePathEnableShCmd;
 
-  if (!swiftBinPath || !fs.existsSync(swiftBinPath)) {
-    window.showErrorMessage(
-      'missing dependent `swift` tool, please configure correct "swift.path.swift_driver_bin"'
-    );
-  }
-  if (!sourcekitePathEnableShCmd) {
-    if (!skProtocolProcess || !fs.existsSync(skProtocolProcess)) {
-      window.showErrorMessage(
-        'missing dependent sourcekite tool, please configure correct "swift.path.sourcekite"'
-      );
-    }
-  }
-  if (!shellPath || !fs.existsSync(shellPath)) {
-    window.showErrorMessage(
-      'missing dependent shell tool, please configure correct "swift.path.shell"'
-    );
-  }
+  // if (!swiftBinPath || !fs.existsSync(swiftBinPath)) {
+  //   window.showErrorMessage(
+  //     'missing dependent `swift` tool, please configure correct "swift.path.swift_driver_bin"'
+  //   );
+  // }
+  // if (!sourcekitePathEnableShCmd) {
+  //   if (!skProtocolProcess || !fs.existsSync(skProtocolProcess)) {
+  //     const action = await window.showErrorMessage(
+  //       `\`sourcekite\` not found at \`${skProtocolProcess}\`.
+  //       Install it and provide the path to \`swift.path.sourcekite\`.`,
+  //       'Retry', 'Help'
+  //     );
+  //     if (action === "Help") {
+  //       await commands.executeCommand(
+  //         "vscode.open",
+  //         Uri.parse("https://github.com/vknabel/vscode-swift-development-environment/tree/2.11.1#using-sourcekite")
+  //       );
+  //     }
+  //   }
+  // }
+  // if (!shellPath || !fs.existsSync(shellPath)) {
+  //   window.showErrorMessage(
+  //     'missing dependent shell tool, please configure correct "swift.path.shell"'
+  //   );
+  // }
 }